Move task and task list classes to ittybotty.data package
