+++ src/main/java/ittybotty/SaveFileManager.java
@@ -1,5 +1,11 @@
 package ittybotty;
 
+import ittybotty.data.TaskList;
+import ittybotty.data.tasks.Event;
+import ittybotty.data.tasks.Task;
+import ittybotty.data.tasks.TaskWithDeadline;
+import ittybotty.data.tasks.ToDo;
+
 import java.io.File;
 import java.io.FileWriter;
 import java.io.IOException;
@@ -69,7 +75,7 @@ public class SaveFileManager {
             case "E" -> new Event(taskInfo.get(2),
                     LocalDate.parse(taskInfo.get(3)),
                     LocalDate.parse(taskInfo.get(4)));
-            default -> throw new IOException("ittybotty.Task type info corrupted.");
+            default -> throw new IOException("ittybotty.data.tasks.Task type info corrupted.");
         };
         if (Boolean.parseBoolean(taskInfo.get(1))) {
             // TODO: handle corruption when neither true nor false
