+++ src/main/java/ittybotty/data/tasks/ToDo.java
@@ -1,4 +1,4 @@
-package ittybotty;
+package ittybotty.data.tasks;
 
 public class ToDo extends Task {
     public ToDo(String name) {
